Unified download functionality
+added tests
+ moved bnci specific code to bnci.py
