+++ moabb/datasets/alex_mi.py
@@ -4,14 +4,12 @@ Alex Motor imagery dataset.
 
 from .base import BaseDataset
 from mne.io import Raw
+import os
 
-from mne.datasets.utils import _get_path, _do_path_update
-from mne.utils import _fetch_file, _url_to_local_path, verbose
+import moabb.datasets.download as dl
 
 ALEX_URL = 'https://zenodo.org/record/806023/files/'
 
-
-@verbose
 def data_path(subject, path=None, force_update=False, update_path=None,
               verbose=None):
     """Get path to local copy of ALEX dataset URL.
@@ -41,27 +39,13 @@ def data_path(subject, path=None, force_update=False, update_path=None,
         Local path to the given data file. This path is contained inside a list
         of length one, for compatibility.
     """  # noqa: E501
-    key = 'MNE_DATASETS_ALEXEEG_PATH'
-    name = 'ALEX'
-    path = _get_path(path, key, name)
     if subject < 1 or subject > 8:
         raise ValueError("Valid subjects between 1 and 8, subject {:d} requested".format(subject))
     url = '{:s}subject{:d}.raw.fif'.format(ALEX_URL, subject)
 
-    destination = _url_to_local_path(url, os.path.join(path, 'MNE-alexeeg-data'))
 
-    # Fetch the file
-    if not os.path.isfile(destination) or force_update:
-        if os.path.isfile(destination):
-            os.remove(destination)
-        if not os.path.isdir(os.path.dirname(destination)):
-            os.makedirs(os.path.dirname(destination))
-        _fetch_file(url, destination, print_destination=False)
-
-    # Offer to update the path
-    _do_path_update(path, update_path, key, name)
-    return destination
-
+    return dl.data_path(url, 'ALEXEEG', path, force_update, update_path, verbose)
+    
 class AlexMI(BaseDataset):
     """Alex Motor Imagery dataset"""
